Import chromium-browser_45.0.2454.85.orig.tar.xz from Ubuntu Vivid
All .gitignore files dropped, as this would have prevented much of the
tarball being imported.
 # find . -name .gitignore -delete

Removed files over 50mb in size to make github happy.
 # find . -type f -size +49M -delete

+++ extensions/browser/api/sockets_udp/udp_socket_event_dispatcher.h
@@ -5,6 +5,8 @@
 #ifndef EXTENSIONS_BROWSER_API_SOCKETS_UDP_UDP_SOCKET_EVENT_DISPATCHER_H_
 #define EXTENSIONS_BROWSER_API_SOCKETS_UDP_UDP_SOCKET_EVENT_DISPATCHER_H_
 
+#include <string>
+
 #include "extensions/browser/api/api_resource_manager.h"
 #include "extensions/browser/api/sockets_udp/sockets_udp_api.h"
 
@@ -71,7 +73,7 @@ class UDPSocketEventDispatcher
                               int bytes_read,
                               scoped_refptr<net::IOBuffer> io_buffer,
                               const std::string& address,
-                              int port);
+                              uint16 port);
 
   // Post an extension event from IO to UI thread
   static void PostEvent(const ReceiveParams& params, scoped_ptr<Event> event);
